migrate string fns from uicommon to MiscUtils
and clean up some leaky usings
